+++ routing/__init__.py
@@ -1,0 +1,3 @@
+from routing.api import routing_router
+
+__all__ = ["routing_router"]